Change the azure resource to be bicep based - Changed KeyVault, AppInsights, ServiceBus and AppConfiguration - Remove provisioners for those resources
