+++ playground/bicep/BicepSample.AppHost/Program.cs
@@ -13,8 +13,8 @@ var templ = builder.AddBicepTemplate("test", "test.bicep")
                    .WithParameter("test", parameter)
                    .WithParameter("values", ["one", "two"]);
 
-var kv = builder.AddBicepKeyVault("kv");
-var appConfig = builder.AddBicepAppConfiguration("appConfig").WithParameter("sku", "standard");
+var kv = builder.AddAzureKeyVault("kv");
+var appConfig = builder.AddAzureAppConfiguration("appConfig").WithParameter("sku", "standard");
 var storage = builder.AddAzureBicepAzureStorage("storage");
                     // .UseEmulator();
 
@@ -32,13 +32,13 @@ var cosmosDb = builder.AddBicepCosmosDb("cosmos")
                       // .UseEmulator()
                       .AddDatabase("db3");
 
-var appInsights = builder.AddBicepApplicationInsights("ai");
+var appInsights = builder.AddApplicationInsights("ai");
 
 // Redis takes forever to spin up...
 var redis = builder.AddRedis("redis")
                    .AsAzureRedis();
 
-var serviceBus = builder.AddBicepAzureServiceBus("sb", ["queue1"], ["topic1"]);
+var serviceBus = builder.AddAzureServiceBus("sb", ["queue1"], ["topic1"]);
 
 builder.AddProject<Projects.BicepSample_ApiService>("api")
        .WithReference(sqlServer)
