+++ src/Aspire.Hosting.Azure.Provisioning/AzureProvisionerExtensions.cs
@@ -7,12 +7,8 @@ using Aspire.Hosting.Azure.Data.Cosmos;
 using Aspire.Hosting.Azure.Provisioning;
 using Aspire.Hosting.Lifecycle;
 using Azure.ResourceManager;
-using Azure.ResourceManager.AppConfiguration;
-using Azure.ResourceManager.ApplicationInsights;
 using Azure.ResourceManager.CosmosDB;
-using Azure.ResourceManager.KeyVault;
 using Azure.ResourceManager.Resources;
-using Azure.ResourceManager.ServiceBus;
 using Azure.ResourceManager.Sql;
 using Azure.ResourceManager.Storage;
 using Microsoft.Extensions.DependencyInjection;
@@ -36,18 +32,8 @@ public static class AzureProvisionerExtensions
         builder.Services.AddOptions<AzureProvisionerOptions>()
             .BindConfiguration("Azure");
 
-        // We're adding 2 because there's no easy way to enumerate all keys and all service types
-        builder.AddAzureProvisioner<AzureKeyVaultResource, KeyVaultProvisioner>();
-        builder.AddResourceEnumerator(resourceGroup => resourceGroup.GetKeyVaults(), resource => resource.Data.Tags);
-
         builder.AddAzureProvisioner<AzureStorageResource, StorageProvisioner>();
         builder.AddResourceEnumerator(resourceGroup => resourceGroup.GetStorageAccounts(), resource => resource.Data.Tags);
-
-        builder.AddAzureProvisioner<AzureServiceBusResource, ServiceBusProvisioner>();
-        builder.AddResourceEnumerator(resourceGroup => resourceGroup.GetServiceBusNamespaces(), resource => resource.Data.Tags);
-
-        builder.AddAzureProvisioner<AzureAppConfigurationResource, AppConfigurationProvisioner>();
-        builder.AddResourceEnumerator(resourceGroup => resourceGroup.GetAppConfigurationStores(), resource => resource.Data.Tags);
 
         builder.AddAzureProvisioner<AzureCosmosDBResource, AzureCosmosDBProvisioner>();
         builder.AddResourceEnumerator(resourceGroup => resourceGroup.GetCosmosDBAccounts(), resource => resource.Data.Tags);
@@ -56,9 +42,6 @@ public static class AzureProvisionerExtensions
         builder.AddResourceEnumerator(resourceGroup => resourceGroup.GetSqlServers(), resource => resource.Data.Tags);
 
         builder.AddAzureProvisioner<AzureBicepResource, BicepProvisioner>();
-
-        builder.AddAzureProvisioner<AzureApplicationInsightsResource, AzureApplicationInsightsProvisioner>();
-        builder.AddResourceEnumerator(resourceGroup => resourceGroup.GetApplicationInsightsComponents(), resource => resource.Data.Tags);
 
         return builder;
     }
