+++ src/Aspire.Hosting.Azure/AzureServiceBusResource.cs
@@ -1,33 +1,30 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Aspire.Hosting.ApplicationModel;
+using Aspire.Hosting.ApplicationModel;
+
+namespace Aspire.Hosting.Azure;
 
 /// <summary>
 /// Represents an Azure Service Bus resource.
 /// </summary>
 /// <param name="name">The name of the resource.</param>
-public class AzureServiceBusResource(string name) : Resource(name), IAzureResource, IResourceWithConnectionString
+public class AzureServiceBusResource(string name) :
+    AzureBicepResource(name, templateResouceName: "Aspire.Hosting.Azure.Bicep.servicebus.bicep"),
+    IResourceWithConnectionString
 {
     /// <summary>
-    /// Gets or sets the Service Bus endpoint. This is the full uri to the service bus 
-    /// namespace, for example <c>"namespace.servicebus.windows.net"</c>.
+    /// Gets the "serviceBusEndpoint" output reference from the bicep template for the Azure Service Bus endpoint.
     /// </summary>
-    public string? ServiceBusEndpoint { get; set; }
+    public BicepOutputReference ServiceBusEndpoint => new("serviceBusEndpoint", this);
 
     /// <summary>
-    /// Gets or sets the names of the queues associated with the Azure Service Bus resource.
+    /// Gets the connection string template for the manifest for the Azure Service Bus endpoint.
     /// </summary>
-    public string[] QueueNames { get; set; } = [];
-    
-    /// <summary>
-    /// Gets or sets the names of the topics associated with the Azure Service Bus resource.
-    /// </summary>
-    public string[] TopicNames { get; set; } = [];
-
+    public string ConnectionStringExpression => ServiceBusEndpoint.ValueExpression;
     /// <summary>
     /// Gets the connection string for the Azure Service Bus endpoint.
     /// </summary>
     /// <returns>The connection string for the Azure Service Bus endpoint.</returns>
-    public string? GetConnectionString() => ServiceBusEndpoint;
+    public string? GetConnectionString() => ServiceBusEndpoint.Value;
 }
