+++ src/Aspire.Hosting.Azure/Extensions/AzureApplicationInsightsExtensions.cs
@@ -1,0 +1,27 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Aspire.Hosting.ApplicationModel;
+using Aspire.Hosting.Azure;
+
+namespace Aspire.Hosting;
+
+/// <summary>
+/// Provides extension methods for adding the Azure ApplicationInsights resources to the application model.
+/// </summary>
+public static class AzureApplicationInsightsExtensions
+{
+    /// <summary>
+    /// Adds an Azure Application Insights resource to the application model.
+    /// </summary>
+    /// <param name="builder">The <see cref="IDistributedApplicationBuilder"/>.</param>
+    /// <param name="name">The name of the resource. This name will be used as the connection string name when referenced in a dependency.</param>
+    /// <returns>A reference to the <see cref="IResourceBuilder{AzureSqlDatabaseResource}"/>.</returns>
+    public static IResourceBuilder<AzureApplicationInsightsResource> AddApplicationInsights(this IDistributedApplicationBuilder builder, string name)
+    {
+        var resource = new AzureApplicationInsightsResource(name);
+        return builder.AddResource(resource)
+                .WithParameter("appInsightsName", resource.CreateBicepResourceName())
+                .WithManifestPublishingCallback(resource.WriteToManifest);
+    }
+}
